moved fixtures to tests
