+++ pkg/appgw/requestroutingrules_test.go
@@ -7,6 +7,7 @@ package appgw
 
 import (
 	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/annotations"
+	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/tests"
 	n "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2018-12-01/network"
 	"github.com/Azure/go-autorest/autorest/to"
 	. "github.com/onsi/ginkgo"
@@ -27,8 +28,8 @@ var _ = Describe("Test SSL Redirect Annotations", func() {
 	expectedRedirectID := agw.redirectConfigurationID(configName)
 
 	Context("test getSslRedirectConfigResourceReference", func() {
-		configBuilder := newConfigBuilderFixture(nil)
-		ingress := newIngressFixture()
+		configBuilder := NewConfigBuilderFixture(nil)
+		ingress := tests.NewIngressFixture()
 
 		actualID := configBuilder.getSslRedirectConfigResourceReference(ingress).ID
 
@@ -38,8 +39,8 @@ var _ = Describe("Test SSL Redirect Annotations", func() {
 	})
 
 	Context("test modifyPathRulesForRedirection with 0 path rules", func() {
-		configBuilder := newConfigBuilderFixture(nil)
-		ingress := newIngressFixture()
+		configBuilder := NewConfigBuilderFixture(nil)
+		ingress := tests.NewIngressFixture()
 		pathMap := newURLPathMap()
 
 		// Ensure there are no path rules defined for this test
@@ -64,8 +65,8 @@ var _ = Describe("Test SSL Redirect Annotations", func() {
 	})
 
 	Context("test modifyPathRulesForRedirection with 1 path rules", func() {
-		configBuilder := newConfigBuilderFixture(nil)
-		ingress := newIngressFixture()
+		configBuilder := NewConfigBuilderFixture(nil)
+		ingress := tests.NewIngressFixture()
 		pathMap := newURLPathMap()
 
 		// Ensure the test is setup correctly
@@ -97,7 +98,7 @@ var _ = Describe("Test SSL Redirect Annotations", func() {
 	})
 
 	Context("test RequestRoutingRules without HTTPS but with SSL Redirect", func() {
-		configBuilder := newConfigBuilderFixture(nil)
+		configBuilder := NewConfigBuilderFixture(nil)
 
 		// TODO(draychev): Move to test fixtures
 		ingress := v1beta1.Ingress{
