+++ pkg/appgw/test_fixtures.go
@@ -6,18 +6,13 @@
 package appgw
 
 import (
-	//"fmt"
-	//"k8s.io/client-go/tools/record"
+	"fmt"
+	"k8s.io/client-go/tools/record"
 
-	//"github.com/Azure/application-gateway-kubernetes-ingress/pkg/annotations"
-	//"github.com/Azure/application-gateway-kubernetes-ingress/pkg/k8scontext"
+	"github.com/Azure/application-gateway-kubernetes-ingress/pkg/k8scontext"
 	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2018-12-01/network"
 	"github.com/Azure/go-autorest/autorest/to"
-	//v1 "k8s.io/api/core/v1"
-	//"k8s.io/api/extensions/v1beta1"
-	//metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	//"k8s.io/apimachinery/pkg/util/intstr"
-	//"k8s.io/client-go/tools/cache"
+	"k8s.io/client-go/tools/cache"
 )
 
 const (
@@ -41,7 +36,7 @@ const (
 	testFixturesSubscription  = "--subscription--"
 )
 
-/* func newAppGwyConfigFixture() network.ApplicationGatewayPropertiesFormat {
+func NewAppGwyConfigFixture() network.ApplicationGatewayPropertiesFormat {
 	feIPConfigs := []network.ApplicationGatewayFrontendIPConfiguration{
 		{
 			// Private IP
@@ -71,9 +66,9 @@ const (
 	return network.ApplicationGatewayPropertiesFormat{
 		FrontendIPConfigurations: &feIPConfigs,
 	}
-} */
+}
 
-/* func newSecretStoreFixture(toAdd *map[string]interface{}) k8scontext.SecretsKeeper {
+func NewSecretStoreFixture(toAdd *map[string]interface{}) k8scontext.SecretsKeeper {
 	c := cache.NewThreadSafeStore(cache.Indexers{}, cache.Indices{})
 	ingressKey := getResourceKey(testFixturesNamespace, testFixturesName)
 	c.Add(ingressKey, testFixturesHost)
@@ -90,13 +85,13 @@ const (
 	return &k8scontext.SecretsStore{
 		Cache: c,
 	}
-} */
+}
 
-/* func keyFunc(obj interface{}) (string, error) {
+func KeyFunc(obj interface{}) (string, error) {
 	return fmt.Sprintf("%s/%s", testFixturesNamespace, testFixturesServiceName), nil
-} */
+}
 
-/* func newConfigBuilderFixture(certs *map[string]interface{}) appGwConfigBuilder {
+func NewConfigBuilderFixture(certs *map[string]interface{}) appGwConfigBuilder {
 	cb := appGwConfigBuilder{
 		appGwIdentifier: Identifier{
 			SubscriptionID: testFixtureSubscription,
@@ -109,10 +104,10 @@ const (
 		backendPoolMap:         make(map[backendIdentifier]*network.ApplicationGatewayBackendAddressPool),
 		k8sContext: &k8scontext.Context{
 			Caches: &k8scontext.CacheCollection{
-				Endpoints: cache.NewStore(keyFunc),
-				Secret:    cache.NewStore(keyFunc),
-				Service:   cache.NewStore(keyFunc),
-				Pods:      cache.NewStore(keyFunc),
+				Endpoints: cache.NewStore(KeyFunc),
+				Secret:    cache.NewStore(KeyFunc),
+				Service:   cache.NewStore(KeyFunc),
+				Pods:      cache.NewStore(KeyFunc),
 			},
 			CertificateSecretStore: NewSecretStoreFixture(certs),
 		},
@@ -121,9 +116,9 @@ const (
 	}
 
 	return cb
-} */
+}
 
-/* func newCertsFixture() map[string]interface{} {
+func NewCertsFixture() map[string]interface{} {
 	toAdd := make(map[string]interface{})
 
 	secretsIdent := secretIdentifier{
@@ -137,237 +132,7 @@ const (
 	toAdd[""] = secretsIdent
 
 	return toAdd
-} */
-
-/* func newIngressBackendFixture(serviceName string, port int32) *v1beta1.IngressBackend {
-	return &v1beta1.IngressBackend{
-		ServiceName: serviceName,
-		ServicePort: intstr.IntOrString{
-			IntVal: port,
-		},
-	}
-} */
-
-/* func newIngressRuleFixture(host string, urlPath string, be v1beta1.IngressBackend) v1beta1.IngressRule {
-	return v1beta1.IngressRule{
-		Host: host,
-		IngressRuleValue: v1beta1.IngressRuleValue{
-			HTTP: &v1beta1.HTTPIngressRuleValue{
-				Paths: []v1beta1.HTTPIngressPath{
-					{
-						Path:    urlPath,
-						Backend: be,
-					},
-				},
-			},
-		},
-	}
-} */
-
-/* func newIngressFixture() *v1beta1.Ingress {
-	be80 := NewIngressBackendFixture(testFixturesServiceName, 80)
-	be443 := NewIngressBackendFixture(testFixturesServiceName, 443)
-
-	return &v1beta1.Ingress{
-		Spec: v1beta1.IngressSpec{
-			Rules: []v1beta1.IngressRule{
-				NewIngressRuleFixture(testFixturesHost, testFixturesURLPath, *be80),
-				NewIngressRuleFixture(testFixturesHost, testFixturesURLPath, *be443),
-			},
-			TLS: []v1beta1.IngressTLS{
-				{
-					Hosts: []string{
-						"www.contoso.com",
-						"ftp.contoso.com",
-						testFixturesHost,
-						"",
-					},
-					SecretName: testFixturesNameOfSecret,
-				},
-				{
-					Hosts:      []string{},
-					SecretName: testFixturesNameOfSecret,
-				},
-			},
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Annotations: map[string]string{
-				annotations.SslRedirectKey: "true",
-			},
-			Namespace: testFixturesNamespace,
-			Name:      testFixturesName,
-		},
-	}
-} */
-
-/* func newServicePortsFixture() *[]v1.ServicePort {
-	httpPort := v1.ServicePort{
-		// The name of this port within the service. This must be a DNS_LABEL.
-		// All ports within a ServiceSpec must have unique names. This maps to
-		// the 'Name' field in EndpointPort objects.
-		// Optional if only one ServicePort is defined on this service.
-		Name: "http",
-
-		// The IP protocol for this port. Supports "TCP", "UDP", and "SCTP".
-		Protocol: v1.ProtocolTCP,
-
-		// The port that will be exposed by this service.
-		Port: 80,
-
-		// Number or name of the port to access on the pods targeted by the service.
-		// Number must be in the range 1 to 65535. Name must be an IANA_SVC_NAME.
-		// If this is a string, it will be looked up as a named port in the
-		// target Pod's container ports. If this is not specified, the value
-		// of the 'port' field is used (an identity map).
-		// This field is ignored for services with clusterIP=None, and should be
-		// omitted or set equal to the 'port' field.
-		TargetPort: intstr.IntOrString{
-			Type:   intstr.Int,
-			IntVal: testFixturesContainerPort,
-		},
-	}
-
-	httpsPort := v1.ServicePort{
-		Name:     "https",
-		Protocol: v1.ProtocolTCP,
-		Port:     443,
-		TargetPort: intstr.IntOrString{
-			Type:   intstr.String,
-			StrVal: "https-port",
-		},
-	}
-
-	randomTCPPort := v1.ServicePort{
-		Name:     "other-tcp-port",
-		Protocol: v1.ProtocolTCP,
-		Port:     554,
-		TargetPort: intstr.IntOrString{
-			IntVal: 8181,
-		},
-	}
-
-	udpPort := v1.ServicePort{
-		Name:     "other-tcp-port",
-		Protocol: v1.ProtocolUDP,
-		Port:     9123,
-		TargetPort: intstr.IntOrString{
-			Type:   intstr.Int,
-			IntVal: 4566,
-		},
-	}
-
-	return &[]v1.ServicePort{
-		httpPort,
-		httpsPort,
-		randomTCPPort,
-		udpPort,
-	}
-}*/
-
-/* func newProbeFixture(containerName string) *v1.Probe {
-	return &v1.Probe{
-		TimeoutSeconds:   5,
-		FailureThreshold: 3,
-		PeriodSeconds:    20,
-		Handler: v1.Handler{
-			HTTPGet: &v1.HTTPGetAction{
-				Host: testFixturesHost,
-				Path: testFixturesURLPath,
-				Port: intstr.IntOrString{
-					Type:   intstr.String,
-					StrVal: containerName,
-				},
-				Scheme: v1.URISchemeHTTP,
-			},
-		},
-	}
-} */
-
-/* func newPodFixture(serviceName string, ingressNamespace string, containerName string, containerPort int32) *v1.Pod {
-	return &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      serviceName,
-			Namespace: ingressNamespace,
-			Labels: map[string]string{
-				testFixturesSelectorKey: testFixturesSelectorValue,
-			},
-		},
-		Spec: v1.PodSpec{
-			Containers: []v1.Container{
-				{
-					Name:  serviceName,
-					Image: "image",
-					Ports: []v1.ContainerPort{
-						{
-							Name:          containerName,
-							ContainerPort: containerPort,
-						},
-					},
-					ReadinessProbe: NewProbeFixture(containerName),
-				},
-			},
-		},
-	}
-} */
-
-/* func newServiceFixture(servicePorts ...v1.ServicePort) *v1.Service {
-	return &v1.Service{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      testFixturesServiceName,
-			Namespace: testFixturesNamespace,
-		},
-		Spec: v1.ServiceSpec{
-			// The list of ports that are exposed by this service.
-			Ports: servicePorts,
-
-			// Route service traffic to pods with label keys and values matching this
-			// selector. If empty or not present, the service is assumed to have an
-			// external process managing its endpoints, which Kubernetes will not
-			// modify. Only applies to types ClusterIP, NodePort, and LoadBalancer.
-			// Ignored if type is ExternalName.
-			Selector: map[string]string{
-				testFixturesSelectorKey: testFixturesSelectorValue,
-			},
-		},
-	}
-} */
-
-/* func newEndpointsFixture() *v1.Endpoints {
-	return &v1.Endpoints{
-		Subsets: []v1.EndpointSubset{
-			{
-				// IP addresses which offer the related ports that are marked as ready. These endpoints
-				// should be considered safe for load balancers and clients to utilize.
-				// +optional
-				Addresses: []v1.EndpointAddress{
-					{
-						IP: "10.9.8.7",
-						// The Hostname of this endpoint
-						// +optional
-						Hostname: "www.contoso.com",
-						// Optional: Node hosting this endpoint. This can be used to determine endpoints local to a node.
-						// +optional
-						NodeName: to.StringPtr(testFixturesNodeName),
-					},
-				},
-				// IP addresses which offer the related ports but are not currently marked as ready
-				// because they have not yet finished starting, have recently failed a readiness check,
-				// or have recently failed a liveness check.
-				// +optional
-				NotReadyAddresses: []v1.EndpointAddress{},
-				// Port numbers available on the related IP addresses.
-				// +optional
-				Ports: []v1.EndpointPort{
-					{
-						Protocol: v1.ProtocolTCP,
-						Name:     testFixturesName,
-						Port:     testFixturesContainerPort,
-					},
-				},
-			},
-		},
-	}
-} */
+}
 
 func newURLPathMap() network.ApplicationGatewayURLPathMap {
 	rule := network.ApplicationGatewayPathRule{
